+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changed
 
+* One can push all items from inventory by using the crate/controller with an empty hand, in addition to sneak-use.
+
 ### Deprecated
 
 ### Removed