+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 
 * Only show crate item in tooltip when looking at the crate front.
+* Sneak+Use opens the create screen from any side.
 
 ### Deprecated